Add activity log endpoint
The created_at timestamp on albums got renamed to published_at and gets
updated when draft gets changed to false now. Images also have publish
dates now which are attached to each of the album associations.

The amount of activitities that are returned by this endpoint will
probably have to to be limited in the future and comments may still need
additional metadata to simplify things for the front end.